tearing everything down and crying
